- Changed NOTIF-NO-CONTENT error to NOTIF-NO-MESSAGE, since a message must ALWAYS be present, otherwise the notification banner in mobile devices won’t show up.
- Added id checking within dialog-closed and dialog-opened events:
since ALL dialogs catch those events, you need to check that the dialog
that was opened/closed is in fact the dialog in question, and only then
execute whatever open/close functionality you want.

- Added activeErrors array in errorFactory service, since the dialog is
created inside the service, and hence no scope can be attached to it
(you can call that dialog/service from any scope); this allows errors
to be stacked (though i can’t see when or if that may happen), and
still be handled correctly (allows their onClose/onOpen function to be
run explicitly on them, and not other dialogs that catch events.

- Removed close ‘x’ from error views and added button which calls
closeThisDialog() $scope function, for simplicity/usability. added css
for button.